requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
llama-index
tavily-python
